fix(gates): raise the ext-typescript handle floor to the band top (#9219 shape), root array_from_values' array and pending values, split lower/tests.rs for the file cap
